nodeApp/requirements.txt
eth-brownie==1.20.0
pytest==6.2.5
pytest-xdist==2.2.1
pytest-cov==2.12.1
coverage==5.5
web3==5.25.0


